Boilerplate stuff.
Adding a folder to document progress through Dart learning for future generations, should be the beginning of a Trepik style guide for Dart just as soon as we know enough for it to be. Also added a repo level readme, but since we don't do anything yet... It's mostly empty.