Made it stateful.
This is the end of this particular page. It will be commented out for the next commit, but we will preserve its code. Basically, lal that happened was a button was added to dynamically change the total mileage and Frodo mileage by 10 per click. Nothing fancy, but a good introduction to state in Flutter.