Saving state because I am leaving computer.
Got a list view of treks as defined classes, plus their distances and taglines. Figured out that some components, like ListView, are designed expressly to be scrollable while others are not, so that's how you get scroll. Lastly, working on adding some images to throw some avatars onto the page as my proficiency with Flutter begins to increase.