Here we go with some real life app construction.
New folder structure, will need to update documentation before filing PR with a systematized description of how to build the fs from here on out. Pages and components separated, pages navigable, and navigation being built. Next step is to extract the app bar and embed it on all the components to have the exact same nav stuff.