A navbar that actually navigates.
Got the appbar setup for bottom navigation, and can actually navigate between pages. Trouble seems to stem now from 2 things: first, the navbar is stateful but is being passed the active tab as a param, which seems like a dumb idea, and second, it needs to be added afresh on every single page, which also seems like a dumb idea. I think what would be eons better is if we could move it to the main component and use its statefulness to paginate, but who knows.