New FS.
The screens dir now contains our file structure. Each screen directory contains the main view as an exportable dart file, and a subdirectory of components to be used on that main dart file. a wrapper dart file sits atop all of it that will essentially be used to determine app state/app configuration variables.