Added anonymous auth.
Good news. The Firebase integration actually works. Don't know really what the point of anonymous auth is. It seems like it could be a good way to track user engagement for users who decide to skip the signup. That adds some overhead to converting those same users to registrations if that person decides to sign up, plus eliminating anonymous users periodically, but it's not the worst idea.