Committing because it's the end of the night.
So we've got the provider package in the spec, we're using it for the StreamProvider to make sure the authentication service can chagne the auth state for the entire app. It was actually really easy, but I'm tired so I'm making silly mistakes, which means it's time to push for the day and call it quits.